FoodSnap/requirements.txt
gradio
transformers
torch
Pillow
requests
